moved toolbar positioning into subfolder
